updated linux bootstrapper

+++ bootstrap_linux.sh
@@ -1,17 +1,13 @@
 #!/bin/bash
 
 echo "symlink .bash_profile"
-rm -f ~/.bash_profile
-ln -s ~/.dotfiles/.bash_profile ~/.bash_profile
+[ ! -f ~/.bash_profile ] && ln -s ~/.dotfiles/.bash_profile ~/.bash_profile
 
 echo "symlink .screenrc"
-rm -f ~/.screenrc
-ln -s ~/.dotfiles/.screenrc ~/.screenrc
+[ ! -f ~/.screenrc ] && ln -s ~/.dotfiles/.screenrc ~/.screenrc
 
 echo "symlink .gitconfig"
-rm -f ~/.gitconfig
-ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
+[ ! -f ~/.gitconfig ] && ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
 
 echo "symlink .gitignore"
-rm -f ~/.gitignore
-ln -s ~/.dotfiles/.gitignore ~/.gitignore
+[ ! -f ~/.gitignore ] && ln -s ~/.dotfiles/.gitignore ~/.gitignore
